Move KubernetesTaskMetadata to its own file
